# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: DealFlow360 API
  description: Comprehensive quotation-to-cash platform API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
paths:
...
  /subscriptions/{subscriptionId}:
    get:
      summary: Get subscription
      tags:
        - Subscriptions
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Subscription details
  /subscriptions/{subscriptionId}/preview-change:
    post:
      summary: Preview subscription change
      tags:
        - Subscriptions
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                planId:
                  type: string
      responses:
        '200':
          description: Change preview
  /subscriptions/{subscriptionId}/change:
    post:
      summary: Change subscription
      tags:
        - Subscriptions
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: number
                planId:
                  type: string
                effectiveDate:
                  type: string
                  format: date-time
      responses:
        '200':
          description: Subscription changed
  /subscriptions/{subscriptionId}/cancel:
    post:
      summary: Cancel subscription
      tags:
        - Subscriptions
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                effectiveDate:
                  type: string
                  format: date-time
                reason:
                  type: string
      responses:
        '200':
          description: Subscription cancelled
  /subscriptions/{subscriptionId}/schedules:
    get:
      summary: List billing schedules
      tags:
        - Subscriptions
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Billing schedules
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session